+++ CHANGELOG.md
@@ -2,7 +2,6 @@
 
 ## [0.2.0](https://github.com/clarity-llm-lang/clarity-agent-cli/compare/clarity-agent-cli-v0.1.0...clarity-agent-cli-v0.2.0) (2026-02-24)
 
-
 ### Features
 
-* bootstrap clarity agent cli repository ([1bad4d9](https://github.com/clarity-llm-lang/clarity-agent-cli/commit/1bad4d9b7032ca1e589aa2a68df8d73d32f40c75))
+- bootstrap clarity agent cli repository ([1bad4d9](https://github.com/clarity-llm-lang/clarity-agent-cli/commit/1bad4d9b7032ca1e589aa2a68df8d73d32f40c75))